crm_project/requirements.txt
Django>=4.2.0
django-environ>=0.11.2
django-cors-headers>=4.3.0
djangorestframework>=3.14.0
django-filter>=23.5
django-crispy-forms>=2.1
crispy-bootstrap5>=2023.10
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0
ib_insync>=0.9.86
celery>=5.3.6
redis>=5.0.1 